requirements.txt
# Data Processing
pandas==2.0.3
numpy==1.24.3

# Machine Learning
scikit-learn==1.3.0
joblib==1.3.1

# API Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# CORS and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.4.2

# Data Validation
email-validator==2.1.0

# Utilities
requests==2.31.0

# Development
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0